requirements.txt
transformers==4.21.1
torch==1.12.0
python-Levenshtein==0.25.1
